+++ research/left_encode-right_encode-benchmark.cpp
@@ -26,7 +26,7 @@
 static std::vector<std::byte>
 left_encode_1(const std::unsigned_integral auto x)
 {
-    const auto w = static_cast<uint8_t>(byte_width(x));
+    const auto w = byte_width(x);
 
     std::vector<std::byte> result;
     result.reserve(1 + w);
@@ -45,7 +45,7 @@ left_encode_1(const std::unsigned_integral auto x)
 static std::vector<std::byte>
 left_encode_2(std::unsigned_integral auto x)
 {
-    const auto w = static_cast<uint8_t>(byte_width(x));
+    const auto w = byte_width(x);
 
     std::vector<std::byte> result;
     result.reserve(1 + w);
@@ -66,7 +66,7 @@ left_encode_2(std::unsigned_integral auto x)
 static std::vector<std::byte>
 left_encode_3(const std::unsigned_integral auto x)
 {
-    const auto w = static_cast<uint8_t>(byte_width(x));
+    const auto w = byte_width(x);
 
     const auto byte_sp = as_byte_span(x);
 
@@ -82,7 +82,7 @@ left_encode_4(std::unsigned_integral auto x)
 {
     fixed_vector<std::byte, 1 + sizeof(decltype(x))> result;
 
-    const auto w = static_cast<uint8_t>(byte_width(x));
+    const auto w = byte_width(x);
 
     result.unchecked_push_back(static_cast<std::byte>(w));
 
@@ -102,7 +102,7 @@ left_encode_5(const std::unsigned_integral auto x)
 {
     fixed_vector<std::byte, 1 + sizeof(decltype(x))> result;
 
-    const auto w = static_cast<uint8_t>(byte_width(x));
+    const auto w = byte_width(x);
 
     result.unchecked_push_back(static_cast<std::byte>(w));
 
@@ -118,7 +118,7 @@ left_encode_5(const std::unsigned_integral auto x)
 static std::vector<std::byte>
 right_encode_1(const std::unsigned_integral auto x)
 {
-    const auto w = static_cast<uint8_t>(byte_width(x));
+    const auto w = byte_width(x);
 
     std::vector<std::byte> result;
     result.reserve(1 + w);
@@ -137,7 +137,7 @@ right_encode_1(const std::unsigned_integral auto x)
 static std::vector<std::byte>
 right_encode_2(std::unsigned_integral auto x)
 {
-    const auto w = static_cast<uint8_t>(byte_width(x));
+    const auto w = byte_width(x);
 
     std::vector<std::byte> result;
     result.reserve(1 + w);
@@ -158,7 +158,7 @@ right_encode_2(std::unsigned_integral auto x)
 static std::vector<std::byte>
 right_encode_3(const std::unsigned_integral auto x)
 {
-    const auto w = static_cast<uint8_t>(byte_width(x));
+    const auto w = byte_width(x);
 
     const auto byte_sp = as_byte_span(x);
 
@@ -175,7 +175,7 @@ right_encode_4(std::unsigned_integral auto x)
 {
     fixed_vector<std::byte, 1 + sizeof(decltype(x))> result;
 
-    const auto w = static_cast<uint8_t>(byte_width(x));
+    const auto w = byte_width(x);
 
     // extract w bytes, going from least significant byte to most significant byte
     for (std::remove_cv_t<decltype(w)> i = 0; i < w; ++i)
@@ -195,7 +195,7 @@ right_encode_5(const std::unsigned_integral auto x)
 {
     fixed_vector<std::byte, 1 + sizeof(decltype(x))> result;
 
-    const auto w = static_cast<uint8_t>(byte_width(x));
+    const auto w = byte_width(x);
 
     const auto byte_sp = as_byte_span(x);
